+++ pendulum/locales/fa/custom.py
@@ -1,0 +1,19 @@
+"""
+fa custom locale file.
+"""
+
+translations = {
+    # Relative time
+    'after': '{0} پس از',
+    'before': '{0} پیش از',
+
+    # Date formats
+    'date_formats': {
+        'LTS': 'HH:mm:ss',
+        'LT': 'HH:mm',
+        'LLLL': 'dddd, D MMMM YYYY HH:mm',
+        'LLL': 'D MMMM YYYY HH:mm',
+        'LL': 'D MMMM YYYY',
+        'L': 'DD/MM/YYYY',
+    },
+}